IKASAN-1483 - Update sample-builder-patter to use @SpringBootApplication and SpringBootTest
