+++ ikasaneip/sample/spring-boot/builder-pattern/src/main/java/com/ikasan/sample/spring/boot/builderpattern/MyApplication.java
@@ -40,16 +40,8 @@
  */
 package com.ikasan.sample.spring.boot.builderpattern;
 
-import org.ikasan.builder.*;
-import org.ikasan.builder.component.ComponentBuilder;
-import org.ikasan.spec.component.endpoint.Consumer;
-import org.ikasan.spec.component.endpoint.EndpointException;
-import org.ikasan.spec.component.endpoint.Producer;
-import org.ikasan.spec.flow.Flow;
-import org.ikasan.spec.module.Module;
-import org.ikasan.spec.resubmission.ResubmissionService;
+import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 /**
  * Sample standalone bootstrap application using the builder pattern.
@@ -57,83 +49,11 @@ import org.springframework.context.annotation.ComponentScan;
  * @author Ikasan Development Team
  */
 @SpringBootApplication
-@ComponentScan(basePackages = {"org.ikasan","com.ikasan"})
+
 public class MyApplication
 {
-    public static void main(String[] args) throws Exception
-    {
-        new MyApplication().executeIM(args);
+    public static void main(String[] args) {
+        SpringApplication.run(MyApplication.class, args);
     }
 
-
-    public IkasanApplication executeIM(String[] args)
-    {
-        // get an ikasanApplication instance
-        IkasanApplication ikasanApplication = IkasanApplicationFactory.getIkasanApplication(MyApplication.class,args);
-
-        // get a builderFactory
-        BuilderFactory builderFactory = ikasanApplication.getBuilderFactory();
-
-        // get a module builder from the ikasanApplication
-        ModuleBuilder moduleBuilder = builderFactory.getModuleBuilder("sample-builder-pattern").withDescription("Example module with pattern builder");
-
-        // get an instance of flowBuilder from the moduleBuilder and create a flow
-        Flow scheduledFlow = getScheduledFlow(moduleBuilder, builderFactory.getComponentBuilder());
-
-        // get an instance of flowBuilder from the moduleBuilder and create a flow
-        Flow jmsFlow = getJmsFlow(moduleBuilder, builderFactory.getComponentBuilder());
-
-        // add flows to the module
-        Module module = moduleBuilder.addFlow(scheduledFlow).addFlow(jmsFlow).build();
-
-        // pass the module to Ikasan to run
-        ikasanApplication.run(module);
-
-        return ikasanApplication;
-    }
-
-    public Flow getScheduledFlow(ModuleBuilder moduleBuilder, ComponentBuilder componentBuilder)
-    {
-        FlowBuilder flowBuilder = moduleBuilder.getFlowBuilder("Scheduled Flow");
-        return flowBuilder.withDescription("scheduled flow description")
-                .consumer("consumer", componentBuilder.scheduledConsumer().setCronExpression("0/5 * * * * ?").setConfiguredResourceId("configuredResourceId")
-                        .setScheduledJobGroupName("scheduledJobGroupName").setScheduledJobName("scheduledJobName").build())
-                .producer("producer", new MyProducer()).build();
-    }
-
-    public Flow getJmsFlow(ModuleBuilder moduleBuilder,ComponentBuilder componentBuilder) {
-        FlowBuilder flowBuilder = moduleBuilder.getFlowBuilder("Jms Flow");
-
-        return flowBuilder.withDescription("Jms flow description")
-                .consumer("consumer", componentBuilder.jmsConsumer().setConfiguredResourceId("configuredResourceId")
-                        .setDestinationJndiName("dynamicQueues/source")
-                        .setConnectionFactoryName("ConnectionFactory")
-                        .setConnectionFactoryJndiPropertyFactoryInitial("org.apache.activemq.jndi.ActiveMQInitialContextFactory")
-                        .setConnectionFactoryJndiPropertyProviderUrl("failover:(vm://embedded-broker?create=false)")
-                        .setDestinationJndiPropertyFactoryInitial("org.apache.activemq.jndi.ActiveMQInitialContextFactory")
-                        .setDestinationJndiPropertyProviderUrl("failover:(vm://embedded-broker?create=false)")
-                        .setAutoContentConversion(true)
-                        .build()
-                )
-                .producer("producer", componentBuilder.jmsProducer()
-                        .setConfiguredResourceId("crid")
-                        .setDestinationJndiName("dynamicQueues/target")
-                        .setConnectionFactoryName("ConnectionFactory")
-                        .setConnectionFactoryJndiPropertyFactoryInitial("org.apache.activemq.jndi.ActiveMQInitialContextFactory")
-                        .setConnectionFactoryJndiPropertyProviderUrl("failover:(vm://embedded-broker?create=false)")
-                        .setDestinationJndiPropertyFactoryInitial("org.apache.activemq.jndi.ActiveMQInitialContextFactory")
-                        .setDestinationJndiPropertyProviderUrl("failover:(vm://embedded-broker?create=false)").build()
-                )
-                .build();
-
-    }
-
-    private class MyProducer implements Producer
-    {
-
-        @Override
-        public void invoke(Object payload) throws EndpointException {
-
-        }
-    }
 }
