+++ ikasaneip/sample/spring-boot/builder-pattern/src/test/java/com/ikasan/sample/spring/boot/builderpattern/ModuleTestConfig.java
@@ -1,10 +1,12 @@
 package com.ikasan.sample.spring.boot.builderpattern;
 
 import org.apache.activemq.ActiveMQConnectionFactory;
+import org.apache.activemq.junit.EmbeddedActiveMQBroker;
 import org.ikasan.testharness.flow.jms.MessageListenerVerifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.jms.config.JmsListenerEndpointRegistry;
 import org.springframework.jms.core.JmsTemplate;
 
@@ -26,10 +28,20 @@ public class ModuleTestConfig
     }
 
     @Bean
+    @DependsOn("broker")
     MessageListenerVerifier messageListenerVerifierTarget()
     {
 
         final MessageListenerVerifier messageListenerVerifier = new MessageListenerVerifier(brokerUrl, "target", registry);
         return messageListenerVerifier;
     }
+
+    @Bean
+    EmbeddedActiveMQBroker broker()
+    {
+        EmbeddedActiveMQBroker broker =  new EmbeddedActiveMQBroker();
+        broker.start();
+        return broker;
+    }
+
 }
